Separate i2c from jshardware.c - still arduino lib
