+++ targets/esp32/jshardware.c
@@ -43,8 +43,8 @@
 #include "rom/uart.h"
 #include "driver/gpio.h"
 
+#include "i2c.h"
 #include "esp32-hal-spi.h"
-#include "esp32-hal-i2c.h"
 
 #define FLASH_MAX (4*1024*1024) //4MB
 #define FLASH_PAGE_SHIFT 12 // Shift is much faster than division by 4096 (size of page)
@@ -59,8 +59,6 @@ static char *tagGPIO = "jshardware(GPIO)";
 static spi_t * _spi[VSPI];
 static uint32_t  g_lastSPIRead = (uint32_t)-1;
 
-// Convert an Espruino pin to an ESP32 pin number.
-static gpio_num_t pinToESP32Pin(Pin pin);
 /**
  * Convert a pin id to the corresponding Pin Event id.
  */
@@ -659,72 +657,6 @@ void jshSPISetReceive(IOEventFlags device, bool isReceive) {
   ESP_LOGD(tag,"<< jshSPISetReceive");
 }
 
-
-//===== I2C =====
-
-// Let's get this working with only one device
-i2c_t * i2c=NULL;
-
-/** Set-up I2C master for ESP32, default pins are SCL:21, SDA:22. Only device I2C1 is supported
- *  and only master mode. */
-void jshI2CSetup(IOEventFlags device, JshI2CInfo *info) {
-  if (device != EV_I2C1) {
-    jsError("Only I2C1 supported"); 
-	return;
-  }
-  Pin scl = info->pinSCL != PIN_UNDEFINED ? info->pinSCL : 21;
-  Pin sda = info->pinSDA != PIN_UNDEFINED ? info->pinSDA : 22;
-
-  //jshPinSetState(scl, JSHPINSTATE_I2C);
-  //jshPinSetState(sda, JSHPINSTATE_I2C);
-   
-  int num=1; // Master mode only 0 is slave mode..
- 
-  i2c_err_t err;
-  i2c = i2cInit(num, 0, false);
-  //ESP_LOGE(tag, "jshI2CSetup: Frequency: %d", info->bitrate);
-  err=i2cSetFrequency(i2c, (uint32_t)info->bitrate);
-  if ( err != I2C_ERROR_OK ) {
-    ESP_LOGE(tag, "jshI2CSetup: i2cSetFrequency error: %d", err);
-	return;
-  }
-  err=i2cAttachSDA(i2c, pinToESP32Pin(sda));
-  if ( err != I2C_ERROR_OK ) {
-    ESP_LOGE(tag, "jshI2CSetup: i2cAttachSDA error: %d", err);
-	return;
-  }  
-  err=i2cAttachSCL(i2c, pinToESP32Pin(scl));
-  if ( err != I2C_ERROR_OK ) {
-    ESP_LOGE(tag, "jshI2CSetup: i2cAttachSCL error: %d", err);
-	return;
-  }
-}
-
-void jshI2CWrite(IOEventFlags device,
-  unsigned char address,
-  int nBytes,
-  const unsigned char *data,
-  bool sendStop) {
-// i2cWrite(i2c_t * i2c, uint16_t address, bool addr_10bit, uint8_t * data, uint8_t len, bool sendStop);
-  i2c_err_t err=i2cWrite(i2c,address,false,data,nBytes,sendStop);
-  if ( err != I2C_ERROR_OK ) {
-    ESP_LOGE(tag, "jshI2CSetup: i2cAttachSCL error: %d", err);
-	return;
-  }
-}
-
-void jshI2CRead(IOEventFlags device,
-  unsigned char address,
-  int nBytes,
-  unsigned char *data,
-  bool sendStop) {
-  i2c_err_t err=i2cRead(i2c,address,false,data,nBytes,sendStop);
-  if ( err != I2C_ERROR_OK ) {
-    ESP_LOGE(tag, "jshI2CSetup: i2cAttachSCL error: %d", err);
-	return;
-  }
-}
-
 //===== System time stuff =====
 
 /**
@@ -914,7 +846,7 @@ unsigned int jshSetSystemClock(JsVar *options) {
 /**
  * Convert an Espruino pin id to a native ESP32 pin id.
  */
-static gpio_num_t pinToESP32Pin(Pin pin) {
+gpio_num_t pinToESP32Pin(Pin pin) {
   if ( pin < 40 ) 
 	return pin + GPIO_NUM_0;
   ESP_LOGE(tag, "pinToESP32Pin: Unknown pin: %d", pin);
